Refactor delete account route
